requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-unfold==0.20.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
